pyproject.toml
[project]
name = "promptchad"
version = "0.1.0"
description = "A/B test prompts across multiple AI providers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "flask>=3.0.0",
    "toml>=0.10.0",
]

[project.scripts]
promptchad = "promptchad:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
